projects/sentiment_analyzer/requirements.txt
# Sentiment Analyzer - LangChain 1.0 Dependencies

langchain-core>=0.3.0
langchain-google-genai>=2.0.0
pydantic>=2.0.0